+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 + Fixed Fast Break not allowing you to break the same block after it was replaced
 + Added "Auto Disable" option to Long Jump for when NCP lags you back
 + Exempt Zombie Pigmen from being Kill Aura targets
++ Add "Timer" option to Speed
 ```
 
 ---